requirements.txt
flask
flask-cors
pymongo
bcrypt
numpy
openai
python-dotenv